Move SnapshotArchiveInfo and friends into its own module
