+++ core/src/accounts_hash_verifier.rs
@@ -9,8 +9,8 @@ use rayon::ThreadPool;
 use solana_gossip::cluster_info::{ClusterInfo, MAX_SNAPSHOT_HASHES};
 use solana_runtime::{
     accounts_db,
+    snapshot_archive_info::SnapshotArchiveInfoGetter,
     snapshot_package::{AccountsPackage, AccountsPackagePre, AccountsPackageReceiver},
-    snapshot_utils::SnapshotArchiveInfoGetter,
 };
 use solana_sdk::{clock::Slot, hash::Hash, pubkey::Pubkey};
 use std::collections::{HashMap, HashSet};
