+++ runtime/src/lib.rs
@@ -34,6 +34,7 @@ pub mod rent_collector;
 pub mod secondary_index;
 pub mod serde_snapshot;
 mod shared_buffer_reader;
+pub mod snapshot_archive_info;
 pub mod snapshot_config;
 pub mod snapshot_package;
 pub mod snapshot_runtime_info;
